feat: settings hub-and-spoke UI, UI modernization, theme improvements
- Split the monolithic settings page into a hub-and-spoke architecture
  with category cards (Appearance, Inbox, Compose, PGP, Navigation,
  Notifications, About) and dedicated sub-screens
- Extracted shared settings composables into SettingsComponents.kt
  (SettingsCard, CardDivider, SectionHeader, SettingsToggleRow,
  ThemeSelectorRow, FontSizeRow, BottomSheetPickerRow, etc.)
- Added AnimatedContent transitions between hub and sub-screens
- Removed deprecated notification permission handling in MainActivity
- Added PGPEncryptionManager initialization in ComposeScreen
- Improved pagination and reload handling in InboxScreen
- Enhanced swipeable email items with snap behavior and action icons
- Updated theme colors and Material 3 styling across the app
- Added icon support and UI improvements to ScheduledMessagesScreen
- Refined IMAP setup screen padding and layout
